Split Payloads into multiple files & use receiveClientPayload() instead of receiveJsonStringMap
