+++ common/src/commonMain/kotlin/com/studo/campusqr/common/payloads/AccessManagementPayloads.kt
@@ -1,7 +1,5 @@
 package com.studo.campusqr.common.payloads
 
-import com.studo.campusqr.common.ClientLocation
-import com.studo.campusqr.common.ClientPayload
 import kotlinx.serialization.Serializable
 
 @Serializable
