+++ common/src/commonMain/kotlin/com/studo/campusqr/common/payloads/LocationsPayloads.kt
@@ -1,6 +1,5 @@
 package com.studo.campusqr.common.payloads
 
-import com.studo.campusqr.common.ClientPayload
 import com.studo.campusqr.common.LocationAccessType
 import kotlinx.serialization.Serializable
 
@@ -15,4 +14,13 @@ class CreateOrUpdateLocationData(
   val name: String,
   val accessType: LocationAccessType,
   val seatCount: Int?
+) : ClientPayload
+
+@Serializable
+class ClientLocation(
+  val id: String,
+  val name: String,
+  var checkInCount: Int?,
+  val accessType: LocationAccessType,
+  val seatCount: Int?,
 ) : ClientPayload
