+++ common/src/commonMain/kotlin/com/studo/campusqr/common/payloads/ReportPayloads.kt
@@ -1,6 +1,5 @@
 package com.studo.campusqr.common.payloads
 
-import com.studo.campusqr.common.ClientPayload
 import kotlinx.serialization.Serializable
 
 @Serializable
