+++ common/src/commonMain/kotlin/com/studo/campusqr/common/payloads/UsersPayloads.kt
@@ -1,6 +1,6 @@
 package com.studo.campusqr.common.payloads
 
-import com.studo.campusqr.common.ClientPayload
+import com.studo.campusqr.common.UserPermission
 import kotlinx.serialization.Serializable
 
 @Serializable
@@ -23,3 +23,20 @@ class EditUserData(
 class DeleteUserData(
   val userId: String
 ) : ClientPayload
+
+@Serializable
+class ClientUser(
+  val id: String,
+  val email: String,
+  val name: String,
+  val permissions: Set<UserPermission>,
+  val firstLoginDate: String,
+) : ClientPayload
+
+// Keep in sync with BackendUser
+val ClientUser.canEditUsers get() = UserPermission.EDIT_USERS in permissions
+val ClientUser.canEditLocations get() = UserPermission.EDIT_LOCATIONS in permissions
+val ClientUser.canViewCheckIns get() = UserPermission.VIEW_CHECKINS in permissions
+val ClientUser.canEditAnyLocationAccess get() = canEditOwnLocationAccess || canEditAllLocationAccess
+val ClientUser.canEditOwnLocationAccess get() = UserPermission.EDIT_OWN_ACCESS in permissions
+val ClientUser.canEditAllLocationAccess get() = UserPermission.EDIT_ALL_ACCESS in permissions
