+++ moderatorFrontend/src/main/kotlin/app/App.kt
@@ -1,8 +1,9 @@
 package app
 
 import MuiPickersUtilsProvider
-import com.studo.campusqr.common.*
+import com.studo.campusqr.common.UserPermission
 import com.studo.campusqr.common.extensions.emptyToNull
+import com.studo.campusqr.common.payloads.*
 import kotlinext.js.js
 import kotlinx.browser.document
 import kotlinx.browser.window
