+++ moderatorFrontend/src/main/kotlin/app/AppContent.kt
@@ -1,6 +1,6 @@
 package app
 
-import com.studo.campusqr.common.UserData
+import com.studo.campusqr.common.payloads.UserData
 import kotlinext.js.js
 import react.RBuilder
 import react.RComponent
