+++ moderatorFrontend/src/main/kotlin/views/guestCheckIn/guestCheckInOverview/GuestCheckInRow.kt
@@ -2,7 +2,7 @@ package views.guestCheckIn.guestCheckInOverview
 
 import com.studo.campusqr.common.extensions.format
 import com.studo.campusqr.common.payloads.ActiveCheckIn
-import com.studo.campusqr.common.payloads.GuestCheckOutData
+import com.studo.campusqr.common.payloads.CheckOutData
 import kotlinx.browser.window
 import react.RBuilder
 import react.RComponent
@@ -56,7 +56,7 @@ class GuestCheckInRow : RComponent<GuestCheckInRowProps, GuestCheckInRowState>()
               launch {
                 val response = NetworkManager.post<String>(
                   "$apiBase/location/$locationId/checkout",
-                  body = GuestCheckOutData(email = props.config.activeCheckIn.email)
+                  body = CheckOutData(email = props.config.activeCheckIn.email)
                 )
                 if (response == "ok") {
                   props.config.onCheckedOut()
