+++ moderatorFrontend/src/main/kotlin/views/locations/AddLocation.kt
@@ -1,9 +1,9 @@
 package views.locations
 
 import app.GlobalCss
-import com.studo.campusqr.common.ClientLocation
 import com.studo.campusqr.common.LocationAccessType
 import com.studo.campusqr.common.extensions.format
+import com.studo.campusqr.common.payloads.ClientLocation
 import com.studo.campusqr.common.payloads.CreateOrUpdateLocationData
 import kotlinext.js.js
 import org.w3c.dom.events.Event
