+++ moderatorFrontend/src/main/kotlin/views/users/AddUser.kt
@@ -1,15 +1,11 @@
 package views.users
 
 import app.GlobalCss
-import com.studo.campusqr.common.ClientUser
-import com.studo.campusqr.common.UserData
 import com.studo.campusqr.common.UserPermission
-import com.studo.campusqr.common.canEditUsers
 import com.studo.campusqr.common.extensions.emailRegex
 import com.studo.campusqr.common.extensions.emptyToNull
 import com.studo.campusqr.common.extensions.format
-import com.studo.campusqr.common.payloads.EditUserData
-import com.studo.campusqr.common.payloads.NewUserData
+import com.studo.campusqr.common.payloads.*
 import kotlinext.js.js
 import org.w3c.dom.events.Event
 import react.*
@@ -73,11 +69,11 @@ class AddUser(props: AddUserProps) : RComponent<AddUserProps, AddUserState>(prop
     val response = NetworkManager.post<String>(
       url = "$apiBase/user/create",
       body = NewUserData(
-          email = state.userEmailTextFieldValue,
-          password = state.userPasswordTextFieldValue,
-          name = state.userNameTextFieldValue,
-          permissions = state.userPermissions.map { it.name }
-        )
+        email = state.userEmailTextFieldValue,
+        password = state.userPasswordTextFieldValue,
+        name = state.userNameTextFieldValue,
+        permissions = state.userPermissions.map { it.name }
+      )
     )
     setState {
       userCreationInProgress = false
@@ -90,13 +86,13 @@ class AddUser(props: AddUserProps) : RComponent<AddUserProps, AddUserState>(prop
     val response = NetworkManager.post<String>(
       url = "$apiBase/user/edit",
       body = EditUserData(
-          userId = (props.config as Config.Edit).user.id,
-          name = state.userNameTextFieldValue.emptyToNull(),
-          password = state.userPasswordTextFieldValue.emptyToNull(),
-          permissions = state.userPermissions
-            .takeIf { (props.config as Config.Edit).user.permissions != it }
-            ?.map { it.name }
-        )
+        userId = (props.config as Config.Edit).user.id,
+        name = state.userNameTextFieldValue.emptyToNull(),
+        password = state.userPasswordTextFieldValue.emptyToNull(),
+        permissions = state.userPermissions
+          .takeIf { (props.config as Config.Edit).user.permissions != it }
+          ?.map { it.name }
+      )
     )
     setState {
       userCreationInProgress = false
